PermissionsIdmResourceTest test is extracted
