+++ src/test/java/gov/ca/cwds/idm/FirstIdmResourceTest.java
@@ -18,7 +18,6 @@ import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_AND_DB_D
 import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_ID;
 import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertExtensible;
 import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertNonStrict;
-import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertStrict;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertTrue;
@@ -42,47 +41,6 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 public class FirstIdmResourceTest extends IdmResourceTest {
-
-  @Test
-  @WithMockCustomUser
-  public void testGetPermissions() throws Exception {
-    assertGetPermissionsSuccess();
-  }
-
-  private void assertGetPermissionsSuccess() throws Exception {
-    MvcResult result =
-        mockMvc
-            .perform(MockMvcRequestBuilders.get("/idm/permissions"))
-            .andExpect(MockMvcResultMatchers.status().isOk())
-            .andExpect(MockMvcResultMatchers.content().contentType(JSON_CONTENT_TYPE))
-            .andReturn();
-
-    assertStrict(result, "fixtures/idm/permissions/valid.json");
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {"OtherRole"})
-  public void testGetPermissionsWithOtherRole() throws Exception {
-    assertGetPermissionsUnauthorized();
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {STATE_ADMIN})
-  public void testGetPermissionsStateAdmin() throws Exception {
-    assertGetPermissionsSuccess();
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {CALS_ADMIN})
-  public void testGetPermissionsCalsAdmin() throws Exception {
-    assertGetPermissionsSuccess();
-  }
-
-  @Test
-  @WithMockCustomUser(roles = {OFFICE_ADMIN})
-  public void testGetPermissionsOfficeAdmin() throws Exception {
-    assertGetPermissionsSuccess();
-  }
 
   @Test
   @WithMockCustomUser
